Изменено название проекта с тестами
Изменено название проекта с тестами на BuildRevisionCounter.Test
