+++ BuildRevisionCounter.Test/App.config
