+++ BuildRevisionCounter.Test/BuildRevisionCounter.Test.csproj
