+++ BuildRevisionCounter.Test/Properties/AssemblyInfo.cs
