+++ BuildRevisionCounter.Test/packages.config
