+++ BuildRevisionCounter.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 12.0.21005.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildRevisionCounter", "BuildRevisionCounter\BuildRevisionCounter.csproj", "{64EEEA30-FCE1-4AA6-A375-EBD721914708}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildRevisionCounterTest", "BuildRevisionCounterTest\BuildRevisionCounterTest.csproj", "{20AE3C20-9E21-43F7-B905-3365AABCEA08}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildRevisionCounter.Test", "BuildRevisionCounter.Test\BuildRevisionCounter.Test.csproj", "{20AE3C20-9E21-43F7-B905-3365AABCEA08}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
